requirements.txt
numpy == 1.16.4
flask == 1.1.1
sqlalchemy==1.2.6
pandas == 0.23.4
cx_Oracle==6.2.1
jinja2==2.10.3
marshmallow==3.3.0
Flask-Bootstrap4==4.0.2